refactor: extract code to a Header component
-Extract TSX code from index.tsx and create Header.tsx component
