+++ src/pages/index.tsx
@@ -5,61 +5,17 @@ import Head from 'next/head';
 import Image from 'next/image';
 import Link from '../components/Link';
 import SpotifyIcon from '../components/SpotifyIcon';
-import GuardBotLogo from '../components/GuardBotLogo';
+import Header from '../components/Header';
 import Footer from '../components/Footer';
 import KeyboardArrowDownOutlinedIcon from '@mui/icons-material/KeyboardArrowDownOutlined';
 
 const Home: NextPage = () => {
-    const menuTabs = ['About', 'The app', 'Other projects'];
-
     return (
         <Box sx={{ backgroundColor: 'background.paper' }}>
             <Head>
                 <title>Spotify Playlist Guard</title>
             </Head>
-            <Box
-                display="flex"
-                alignItems="center"
-                padding="1rem 1rem"
-                sx={{ backgroundColor: '#000000' }}
-            >
-                <Box display="flex-columns">
-                    <Link
-                        href="/"
-                        sx={{ textDecoration: 'none' }}
-                        display="flex"
-                    >
-                        <GuardBotLogo height="30px" width="30px" />
-                    </Link>
-                </Box>
-                <Box
-                    display="flex"
-                    position="relative"
-                    left="50%"
-                    sx={{ transform: 'translateX(-50%)' }}
-                >
-                    {menuTabs.map((menuTab) => {
-                        return (
-                            <Link
-                                key={menuTab}
-                                href="/"
-                                sx={{ textDecoration: 'none' }}
-                                margin="0 2rem"
-                                underline="hover"
-                                color="text.primary"
-                            >
-                                <Typography
-                                    variant="body2"
-                                    align="center"
-                                    color={'text.primary'}
-                                >
-                                    {menuTab}
-                                </Typography>
-                            </Link>
-                        );
-                    })}
-                </Box>
-            </Box>
+            <Header />
             <Box
                 display="flex"
                 flexDirection="column"
